refactor(canvas): centralize placement algorithm in placement.ts
Move the whole recommendation algorithm — recommendPlacements, findFreePosition,
nudgeToFree, the geometry helpers (clustering, empty-rectangle finding, gap-fill,
clamping) and the placement constants/types — out of the 2k-line canvas store
into one focused, pure, dependency-free module (canvas/placement.ts). The store
now just imports it and owns the picker state/actions; rendering lives in
GhostPlacementLayer. Drops the dead placementSizeVariants/PlacementSizeVariant.

No behaviour change; tests updated to import from the new module.
